+++ src/costansts.ts
@@ -1,2 +1,3 @@
 export const MIN_CONSOLE_WIDTH = 125;
 export const PIPELINES_PAGE_SIZE = 50; // max 100;
+export const PIPELINES_CHECK_SIZE = 1000;