+++ src/commands/tags.ts
@@ -1,11 +1,13 @@
 import { Project } from '../common/Config';
 import { createTagOnRef, Response, StatusCode } from '../common/api';
 import { PipelineCommand } from '../common/pipelines';
+import { sleep } from '../common/sleep';
 
 export class Tags extends PipelineCommand {
   protected async runPerProject(project: Project) {
     const tag = await this.crateTag(project);
     if (this.yargs.await) {
+      await sleep(5000);
       const pipeline = await this.getPipeline(project, this.yargs.ref);
       return await this.awaitPipelineCompletion(project, pipeline.data);
     }