+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change log format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## Next
 ### Features
 - new aliases for pipeline & deploy
+- increased pipeline lookout to 1000
 
 ## 1.0.2 - 2020-05-15
 ### Features